feat(matrix-output): impl matrix-output-write
