+++ matrix-output-write/package.json
@@ -7,14 +7,14 @@
   "keywords": [],
   "scripts": {
     "start": "run-p build && node dist/index.js",
-    "build": "tsc",
+    "build": "ncc build src/index.ts -o dist",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "type-check": "tsc --noEmit",
     "test": "vitest"
   },
   "dependencies": {
-    "@actions/artifact": "^2.2.1",
+    "@actions/artifact": "^2.2.0",
     "@actions/core": "^1.11.1",
     "tmp": "^0.2.3",
     "typescript": "^5.7.3",
@@ -22,10 +22,11 @@
   },
   "devDependencies": {
     "@shiron-dev/eslint-config": "2.1.0",
-    "@types/node": "22.10.6",
-    "eslint": "9.18.0",
+    "@types/node": "22.10.5",
+    "@vercel/ncc": "^0.38.3",
+    "eslint": "9.17.0",
     "eslint-plugin-format": "1.0.1",
-    "npm-run-all2": "5.0.2",
+    "npm-run-all2": "5.0.0",
     "vitest": "2.1.8"
   }
 }
