--- matrix-output-write/test/action.test.ts
@@ -1,5 +1,0 @@
-import { expect, it } from "vitest";
-
-it("always passes", () => {
-  expect(true).toBe(true);
-});
